refac(selected race): did the same deal with single race, data is now handled through a seperate component
